moving files to the correct place.
